requirements/dev.txt
asgiref==3.3.1
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.3
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-debug-toolbar==3.2
django-embed-video==1.4.8
django-extensions==3.1.3
django-taggit==3.1.0
idna==2.10
oauthlib==3.1.0
phpserialize==1.3
Pillow==8.0.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tzdata==2022.1
urllib3==1.26.2

